Trait finl_unicode::categories::CharacterCategories

pub trait CharacterCategories {

    // Required methods
    fn is_letter(self) -> bool;
    fn is_letter_cased(self) -> bool;
    fn is_letter_uppercase(self) -> bool;
    fn is_letter_lowercase(self) -> bool;
    fn is_letter_titlecase(self) -> bool;
    fn is_letter_modifier(self) -> bool;
    fn is_letter_other(self) -> bool;
    fn is_mark(self) -> bool;
    fn is_mark_nonspacing(self) -> bool;
    fn is_mark_spacing(self) -> bool;
    fn is_mark_enclosing(self) -> bool;
    fn is_letter_or_mark(self) -> bool;
    fn is_number(self) -> bool;
    fn is_number_decimal(self) -> bool;
    fn is_number_letter(self) -> bool;
    fn is_number_other(self) -> bool;
    fn is_punctuation(self) -> bool;
    fn is_punctuation_connector(self) -> bool;
    fn is_punctuation_dash(self) -> bool;
    fn is_punctuation_open(self) -> bool;
    fn is_punctuation_close(self) -> bool;
    fn is_punctuation_initial(self) -> bool;
    fn is_punctuation_final(self) -> bool;
    fn is_punctuation_other(self) -> bool;
    fn is_symbol(self) -> bool;
    fn is_symbol_math(self) -> bool;
    fn is_symbol_currency(self) -> bool;
    fn is_symbol_modifier(self) -> bool;
    fn is_symbol_other(self) -> bool;
    fn is_separator(self) -> bool;
    fn is_separator_space(self) -> bool;
    fn is_separator_line(self) -> bool;
    fn is_separator_paragraph(self) -> bool;
    fn is_other(self) -> bool;
    fn is_control(self) -> bool;
    fn is_format(self) -> bool;
    fn is_private_use(self) -> bool;
    fn is_unassigned(self) -> bool;
    fn get_major_category(self) -> MajorCategory;
    fn get_minor_category(self) -> MinorCategory;

}

Required Methods

fn is_letter(self) -> bool

Determines whether a character is class L, letter (Lu, Ll, Lt, Lm, Lo). This includes all characters used for word formation, both phonetic and ideograms. It does not include symbols that look like letters but are really numbers, such as the encoded Roman numerals.

fn is_letter_cased(self) -> bool

Determines whether a character is class LC, cased letter (Lu, Ll, Lt). This will include letters in the Adlam, Armenian, Cherokee, Coptic, Cyrillic, Deseret, Georgian (Mtavruli), Glagolitic, Greek, Medefaidrin, Old Hungarian, Osage, Roman, Vithkuqi and Warang Citi alphabets (including their extensions) along with special math versions of letters including the Angstrom symbol and Ohm symbol and other such characters like ℇ, ℑ, ℕ, or Ↄ.

fn is_letter_uppercase(self) -> bool

Determines whether a character is an upper-case letter (Lu)

fn is_letter_lowercase(self) -> bool

Determines whether a character is a lower-case letter (Ll)

fn is_letter_titlecase(self) -> bool

Determines whether a character is a title-case letter (Lt). This is composed of Latin digraphs such as ǅ and Greek capital letters with the iota subscript which when uppercased and rendered with prosgegrammeni should be printed as, e.g., ΑΙ but in titlecase would be Αι (although other possibilities for rendering both may occur depending on the typeface.

Inquiry Is there a use for being able to verify that something is Lu or Lt? A bit-mask is possible to avoid the logical or and would be mildly faster.

fn is_letter_modifier(self) -> bool

Determines whether a character is a modifier letter (Lm). These are letters or symbols typically written adjacent to other letters which modify their usage in some way. They are mostly used in phonetic contexts

fn is_letter_other(self) -> bool

Determines whether a character is an other letter (Lo). This is all the characters from uncased scripts alongside a handful of miscellaneous uncased letters such as ª or ƻ.

fn is_mark(self) -> bool

Determines whether a character is a mark (M). These are characters that combine with other characters in various ways (e.g., for accents or other modifiers). A character plus a mark is generally viewed as a single grapheme. Note for most cases you will want to use grapheme segmentation to identify individual graphemes and word segmentation to identify word boundaries rather than examining categories to find word/grapheme boundaries.

fn is_mark_nonspacing(self) -> bool

Determines whether a character is a nonspacing mark (Mn). This is largely diacritics which will not change the width of the character even if they are placed before or after the base character (as is the case with notation for Znamenny chant).

fn is_mark_spacing(self) -> bool

Determines whether a character is a spacing mark (Mc).This is largely used by some scripts derived from the classical Brahmic script where spacing modifiers are added to consonants to indicate vowel sounds and for some musical symbols.

fn is_mark_enclosing(self) -> bool

Determines whether a character is an enclosing mark (Me). This is a small category of combining marks like ㅤ⃠ which will surround their modified character, e.g., z⃠ which may be rendered as a slashed circle over a z.

fn is_letter_or_mark(self) -> bool

Determines if the category is either L or M (this is used by finl to determine whether to continue a sequence for a named command. If you’re tempted to use this, ask whether you really want this or if you’re better off using grapheme or word segmentation or UAX31 identifier and pattern syntax.

fn is_number(self) -> bool

Determines whether a character is a number (N). This includes digits in various scripts along with special representations of numbers like Unicode encodings of Roman numerals or decorated digits like 3︎⃣ and renderings of fractions.

fn is_number_decimal(self) -> bool

Determines whether a character is a decimal digit (Nd). This includes the expected 0–9 digits along with their equivalents in other scripts that use decimal digits, but potentially with other forms plus some special representations of the digits for mathematics and legacy encodings (the last include double-struck digits like 𝟚 and LED-style digits like 🯴 (U+1FBF4).

fn is_number_letter(self) -> bool

Determines whether a character is a letterlike numeric character (Nl). This includes Unicode representation of Roman numerals as well as special characters for non-place value numbering systems in other scripts (e.g., Greek letter-based numbers, Gothic, Old Persian, Cuneiform, etc.)

fn is_number_other(self) -> bool

Determines whether a character is an other numeric character (No). This includes fractions in Hindu-Arabic (Western) and other scripts, super- and subscripted digits and decorated digits along with digits from numbering systems which are not decimal but also not alphabetical in their form like Old Hungarian or Sogdian.

fn is_punctuation(self) -> bool

Determines whether a character is a punctuation character (P)

fn is_punctuation_connector(self) -> bool

Determines whether a character is a connector punctuation character (Pc). These include the underscore character and CJK variations on that character.

fn is_punctuation_dash(self) -> bool

Determines whether a character is a dash punctuation character (Pd). This includes the hyphen, typographical dashes and variations for non-Latin scripts.

fn is_punctuation_open(self) -> bool

Determines whether a character is an open punctuation character (Ps). This includes all the variations on parentheses, braces and brackets in different scripts including mathematical and computational special symbols along with the low-9 single and double quotation markes (‚ and „) but not other quotation marks.

fn is_punctuation_close(self) -> bool

Determines whether a character is a close punctuation character (Pe). This includes all the variations on parentheses, braces and brackets in different scripts including mathematical and computational special symbols.

fn is_punctuation_initial(self) -> bool

Determines whether a character is an initial punctuation character (Pi). This includes variations on opening quotation marks, functionally similar symbols from the New Testament editorial symbols set and the left vertical bar with quill (which is apparently meant to be used to build extended versions of ⁅ if I understand the reference in the Unicode character chart).

fn is_punctuation_final(self) -> bool

Determines whether a character is a final punctuation character (Pf). This includes variations on closing quotation marks, functionally similar symbols from the New Testament editorial symbols set and the left vertical bar with quill (which is apparently meant to be used to build extended versions of ⁆ if I understand the reference in the Unicode character chart).

fn is_punctuation_other(self) -> bool

Determines whether a character is an other punctuation character (Po). This is all the punctuation symbols which are not part of the above set.

fn is_symbol(self) -> bool

Determines whether a character is a symbol (S). These are non-punctuation symbols including currency and mathematical symbols as well as dingbats and emoji.

fn is_symbol_math(self) -> bool

Determines whether a character is a math symbol (Sm).

fn is_symbol_currency(self) -> bool

Determines whether a character is a currency symbol (Sc)

fn is_symbol_modifier(self) -> bool

Determines whether a character is a modifier symbol (Sk). This is mostly stand-alone versions of combining accents, along with the Emoji skintone modifiers.

fn is_symbol_other(self) -> bool

Determines whether a character is an other symbol (So). This is all the symbols which neither have mathematical uses nor are currency symbols.

fn is_separator(self) -> bool

Determines whether a character is a separator (Z). This is all spaces plus the Unicode line and paragraph separators.

fn is_separator_space(self) -> bool

Determines whether a character is a space separator (Zs). In addition to the standard space and the non-breaking space (U+A0), there are various typographic spaces and the Ogham space mark but printed spaces from other scripts such as the Ge’ez word space (፡).

fn is_separator_line(self) -> bool

Determines whether a character is a line separator (Zl). This is the Unicode code point U+2028.

fn is_separator_paragraph(self) -> bool

Determines whether a character is a paragraph separator (Zp). This is the Unicode code point U+2029.

fn is_other(self) -> bool

Determines whether a character is an other character (C). These consist of control characters, format characters, surrogates (no test is provided for these since they will not appear in valid UTF-8 text), unassigned characters and private use characters.

fn is_control(self) -> bool

Determines whether a character is a control character (Cc). These are the 65 characters in the ranges 0x00–0x1f and 0x7f–0x9f.

fn is_format(self) -> bool

Determines whether a character is a format character (Cf). This is a mix of non-printing characters like soft hyphen (U+00AD) and Zero Width Joiner (ZWJ—U+200D) and the tag block which is used in Emoji sequences for flags along with some printing characters for non-Latin script (mostly Arabic but also Mongolian).

fn is_private_use(self) -> bool

Determines whether a character is a private use character (Co)

fn is_unassigned(self) -> bool

Determines whether a character is unassigned (Cn)

fn get_major_category(self) -> MajorCategory

Get the major category for a character (L, M, N, P, S, Z or C). Note that the “special” category of LC cannot be returned by this method since it is not a mutually exclusive category.

fn get_minor_category(self) -> MinorCategory

get the minor category for a character (Lu, Ll, Lt, Lm, Lo, Mn, Mc, Me, Nd, Nl, No, Pc, Pd, Ps, Pi, Pf, Po, Sm, Sk, Sc, So, Za, Zl, Zp, Cc, Cf, Co, Cn). Note that the “special” category of LC cannot be returned by this method since it is not a mutually exclusive category.